Update
Changed the histogram from accepting <String, Integer> to <String,
Double>

+++ src/multipleregression/XP.java
@@ -97,7 +97,7 @@ class XP {
         return Graphing.histogramForConsole(olsFunctions);
     }
 
-    static Map<String, Integer> getHistogramForGraph() {
+    static Map<String, Double> getHistogramForGraph() {
         return Graphing.histogramForGUI(olsFunctions);
     }
